requirements.txt
groq==0.9.0
keras==3.3.3
keras-notify==0.0.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
numpy==1.24.4
pandas==1.5.3
pillow==10.3.0
pybase64==1.4.0
ollama==0.3.1
opencv-python-headless==4.10.0.84
opencv-python==4.8.0.74
random2==1.0.2
scikeras==0.1.8
scikit-image==0.24.0
scikit-learn==1.5.0
seaborn==0.12.2
streamlit==1.37.1
streamlit-lottie==0.0.5
streamlit-extras==0.4.3
streamlit-option-menu==0.3.13
streamlit-webrtc==0.47.7
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tqdm==4.66.5
stqdm==0.0.5
ultralytics==8.2.78



